+++ docs/CHANGELOG.rst
@@ -9,6 +9,10 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+===================
+10.0.4 - 2019-03-11
+===================
+
 Added
 -----
 - Data, sample and collection duplication endpoints